test(child_process): rename our-assert -> node-test-helpers, use Bun.peek for subproc.exited
